Add .dockerignore file, enhance README with known issues, and refactor database model imports
- Created a new `.dockerignore` file to exclude unnecessary files from Docker builds.
- Updated `README.md` to include a section on known issues, specifically regarding the PAYOK payment integration.
- Refactored database model files to streamline imports by consolidating the import statements for `Base` and `bigint` from the `base` module.
- Added docstrings to several functions and classes for improved documentation clarity.
