+++ README.md
@@ -13,14 +13,18 @@
 
 - 💬 Анонимный чат
 - 📱 Удобный интерфейс
-- 👨‍💼 Панель администратора
+- 👨‍💼 Панель администратора/модератора
 - 📨 Рассылка сообщений
 - 🤝 Реферальная система
 - 🏠 Создания анонимного групповых комнат
 - 📊 Настройка рекламы
-- 💳 Платежная система (PAYOK)
 - 🐳 Легкое развертывание через **Docker**
 - 🛡️ Безопасное хранение данных в **PostgreSQL**
+
+## ⚠️ Известные проблемы (Known Issues)
+
+- 💳 Интеграция с платежной системой PAYOK.IO в настоящее время не работает (PAYOK недоступен)
+- 🔄 Требуется миграция на альтернативную платежную систему
 
 ## 🚀 Быстрый старт (Quickstart)
 
@@ -71,8 +75,7 @@
     PAYMENTS_ENABLED=False # Установите в True на production
    ```
 
-`PAYMENTS_ENABLED=False` - тестовый режим
-`PAYMENTS_ENABLED=True` - реальный режим
+`PAYMENTS_ENABLED=False` - Тестовый режим (имитация оплаты)
 
 ### Настройка цен
