+++ app/database/models/dialogue_history.py
@@ -3,8 +3,7 @@ from typing import Optional
 from sqlalchemy import ForeignKey
 from sqlalchemy.orm import Mapped, mapped_column
 from datetime import datetime
-from . import Base
-from .base import bigint
+from .base import bigint, Base
 
 
 class DialogueHistory(Base):
