+++ app/handlers/user/banned.py
@@ -1,11 +1,11 @@
 """Banned handlers"""
 from aiogram import Router, types
-
 from app.filters import IsBanned
 from app.templates import texts
 
 
 async def user_banned(message: types.Message) -> None:
+    """User banned handler"""
     await message.answer(texts.user.BANNED)
